refacto(fuzzing): refacto fuzzing to handle add/sub/mul
