+++ core/tests/templates/operation.sierra
@@ -1,0 +1,17 @@
+type felt = felt;
+
+libfunc felt_const<{lhs}> = felt_const<{lhs}>;
+libfunc felt_const<{rhs}> = felt_const<{rhs}>;
+libfunc store_temp<felt> = store_temp<felt>;
+libfunc felt_{op} = felt_{op};
+libfunc rename<felt> = rename<felt>;
+
+felt_const<{lhs}>() -> ([0]);
+felt_const<{rhs}>() -> ([1]);
+store_temp<felt>([0]) -> ([0]);
+felt_{op}([0], [1]) -> ([2]);
+store_temp<felt>([2]) -> ([2]);
+rename<felt>([2]) -> ([3]);
+return([3]);
+
+add::add::main@0() -> (felt);
