+++ fuzzing/Cargo.toml
